requirements.txt
# БЮДЖЕТНА ВЕРСІЯ - мінімальні залежності
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
django-storages[azure]>=1.13.0
python-decouple>=3.6
whitenoise>=6.0.0
